functions/scrape_prices/main.py
import json
import logging
import os
from datetime import datetime

from google.cloud import storage
from playwright.sync_api import sync_playwright


def scrape_prices(request):
    """Cloud Function to scrape iPhone prices."""
    try:
        # Initialize logging
        logging.basicConfig(level=logging.INFO)
        logger = logging.getLogger(__name__)
        
        # Run the scraping process
        with sync_playwright() as p:
            browser = p.chromium.launch()
            page = browser.new_page()
            
            # Add your scraping logic here
            # This is a placeholder - you'll need to implement the actual scraping logic
            
            browser.close()
        
        return {"status": "success", "message": "Price scraping completed successfully"}
        
    except Exception as e:
        logger.error(f"Error during price scraping: {str(e)}")
        return {"status": "error", "message": str(e)} 