functions/scrape_prices/requirements.txt
google-cloud-storage==2.14.0
playwright==1.42.0 